oops

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: bvaujour <bvaujour@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 12:16:25 by bvaujour          #+#    #+#             */
-/*   Updated: 2023/06/13 11:27:46 by bvaujour         ###   ########.fr       */
+/*   Updated: 2023/06/13 16:10:04 by bvaujour         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	int	i;
